[NF][SECURITY-64] Интерфейс управления правами доступа

See merge request framework/security-admin!225

+++ security-admin-impl/src/main/java/net/n2oapp/security/admin/impl/service/PermissionServiceImpl.java
@@ -1,13 +1,18 @@
 package net.n2oapp.security.admin.impl.service;
 
+import net.n2oapp.platform.i18n.UserException;
 import net.n2oapp.security.admin.api.criteria.PermissionCriteria;
+import net.n2oapp.security.admin.api.model.AppSystem;
 import net.n2oapp.security.admin.api.model.Permission;
+import net.n2oapp.security.admin.api.model.PermissionUpdateForm;
 import net.n2oapp.security.admin.api.service.PermissionService;
 import net.n2oapp.security.admin.impl.entity.PermissionEntity;
 import net.n2oapp.security.admin.impl.entity.SystemEntity;
 import net.n2oapp.security.admin.impl.repository.PermissionRepository;
 import net.n2oapp.security.admin.impl.service.specification.PermissionSpecifications;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
@@ -31,16 +36,28 @@ public class PermissionServiceImpl implements PermissionService {
 
     @Override
     public Permission create(Permission permission) {
+        checkPermissionCodeUnique(permission.getCode());
         return model(permissionRepository.save(entity(permission)));
     }
 
     @Override
-    public Permission update(Permission permission) {
-        return model(permissionRepository.save(entity(permission)));
+    public Permission update(PermissionUpdateForm permission) {
+        PermissionEntity permissionForUpdate = permissionRepository.getOne(permission.getCode());
+        permissionForUpdate.setName(permission.getName());
+        if (permission.getParent() != null && permission.getParent().getCode() != null) {
+            checkParent(permission);
+            permissionForUpdate.setParentPermission(new PermissionEntity(permission.getParent().getCode()));
+        } else
+            permissionForUpdate.setParentPermission(null);
+        return model(permissionRepository.save(permissionForUpdate));
     }
 
     @Override
     public void delete(String code) {
+        PermissionEntity permissionEntity = permissionRepository.findById(code).get();
+        checkRolesWithSuchPermission(permissionEntity);
+        if (permissionEntity.getHasChildren())
+            throw new UserException("exception.permissionHasChildren");
         permissionRepository.deleteById(code);
     }
 
@@ -51,44 +68,39 @@ public class PermissionServiceImpl implements PermissionService {
     }
 
     @Override
-    public List<Permission> getAll(PermissionCriteria criteria) {
+    public Page<Permission> getAll(PermissionCriteria criteria) {
         Specification<PermissionEntity> specification = new PermissionSpecifications(criteria);
         if (criteria.getOrders() == null) {
             criteria.setOrders(Arrays.asList(new Sort.Order(Sort.Direction.ASC, "code")));
         } else {
             criteria.getOrders().add(new Sort.Order(Sort.Direction.ASC, "code"));
         }
-        return permissionRepository.findAll(specification, criteria).stream().map(this::model).collect(Collectors.toList());
+
+        if (criteria.getWithoutParent() != null && criteria.getWithoutParent()) {
+            return getAllWithoutAncestor(specification, criteria);
+        }
+        Page<PermissionEntity> permissionEntityPage = permissionRepository.findAll(specification, criteria);
+        if (criteria.getWithSystem() != null && criteria.getWithSystem()) {
+            return addPermissionFromSystem(permissionEntityPage);
+        }
+        return permissionEntityPage.map(this::model);
     }
 
     @Override
     public List<Permission> getAllByParentCode(String parentCode) {
-        return permissionRepository.findByParentCode(parentCode).stream().map(this::model).collect(Collectors.toList());
+        return permissionRepository.findByParentPermission(new PermissionEntity(parentCode)).stream().map(this::model).collect(Collectors.toList());
     }
 
-    @Override
-    public List<Permission> getAllByParentIdIsNull() {
-        return permissionRepository.findByParentCodeIsNull().stream().map(this::model).collect(Collectors.toList());
-    }
-
-    @Override
-    public List<Permission> getAllWithSystem(PermissionCriteria criteria) {
-        Specification<PermissionEntity> specification = new PermissionSpecifications(criteria);
-        if (criteria.getOrders() == null) {
-            criteria.setOrders(Arrays.asList(new Sort.Order(Sort.Direction.ASC, "code")));
-        } else {
-            criteria.getOrders().add(new Sort.Order(Sort.Direction.ASC, "code"));
-        }
-
-        List<PermissionEntity> permissions = permissionRepository.findAll(specification, criteria).stream().collect(Collectors.toList());
-        Set<SystemEntity> systems = new HashSet<>();
+    private Page<Permission> addPermissionFromSystem(Page<PermissionEntity> permissionEntityPage) {
+        List<PermissionEntity> permissions = permissionEntityPage.stream().collect(Collectors.toList());
         List<Permission> modelPermissions = new ArrayList<>();
+        Set<SystemEntity> systems = new HashSet<>();
         for (int i = 0; i < permissions.size(); i++) {
             PermissionEntity permission = permissions.get(i);
-            if (isNull(permission.getParentCode())
+            if (isNull(permission.getParentPermission())
                     && nonNull(permission.getSystemCode())) {
                 Permission temp = model(permission);
-                temp.setParentCode("$" + permission.getSystemCode().getCode());
+                temp.setParent(new Permission("$" + permission.getSystemCode().getCode()));
                 modelPermissions.add(temp);
                 permissions.remove(permission);
                 i--;
@@ -103,7 +115,24 @@ public class PermissionServiceImpl implements PermissionService {
             }
         }
         modelPermissions.addAll(permissions.stream().map(this::model).collect(Collectors.toList()));
-        return modelPermissions;
+        return new PageImpl<>(modelPermissions);
+    }
+
+    private Page<Permission> getAllWithoutAncestor(Specification<PermissionEntity> specification, PermissionCriteria criteria) {
+        List<PermissionEntity> permissions = permissionRepository.findAll(specification);
+        List<PermissionEntity> checkForRemove = new ArrayList<>();
+        List<PermissionEntity> forRemove = new ArrayList<>();
+        for (PermissionEntity permission : permissions) {
+            checkForRemove.clear();
+            while (permission.getParentPermission() != null) {
+                checkForRemove.add(permission);
+                if (criteria.getCode().equals(permission.getParentPermission().getCode()))
+                    forRemove.addAll(checkForRemove);
+                permission = permission.getParentPermission();
+            }
+        }
+        permissions.removeAll(forRemove);
+        return new PageImpl<>(permissions.stream().map(this::model).collect(Collectors.toList()));
     }
 
     private PermissionEntity entity(Permission model) {
@@ -111,9 +140,11 @@ public class PermissionServiceImpl implements PermissionService {
         PermissionEntity entity = new PermissionEntity();
         entity.setName(model.getName());
         entity.setCode(model.getCode());
-        entity.setParentCode(model.getParentCode());
-        if (model.getSystemCode() != null)
-            entity.setSystemCode(new SystemEntity(model.getSystemCode()));
+        entity.setUserLevel(model.getUserLevel());
+        if (model.getParent() != null && model.getParent().getCode() != null)
+            entity.setParentPermission(new PermissionEntity(model.getParent().getCode()));
+        if (model.getSystem() != null && model.getSystem().getCode() != null)
+            entity.setSystemCode(new SystemEntity(model.getSystem().getCode()));
         return entity;
     }
 
@@ -122,10 +153,52 @@ public class PermissionServiceImpl implements PermissionService {
         Permission model = new Permission();
         model.setName(entity.getName());
         model.setCode(entity.getCode());
-        model.setParentCode(entity.getParentCode());
         model.setHasChildren(entity.getHasChildren());
-        if (entity.getSystemCode() != null)
-            model.setSystemCode(entity.getSystemCode().getCode());
+        model.setUserLevel(entity.getUserLevel());
+        if (entity.getParentPermission() != null) {
+            model.setParent(model(entity.getParentPermission()));
+        }
+        if (entity.getSystemCode() != null) {
+            model.setSystem(new AppSystem(entity.getSystemCode().getCode()));
+            model.getSystem().setName(entity.getSystemCode().getName());
+        }
         return model;
+    }
+
+    /**
+     * Валидация на создание права доступа
+     * Запрещено создавать право доступа, если существует право доступа с таким кодом
+     */
+    private void checkPermissionCodeUnique(String code) {
+        Optional<PermissionEntity> permissionEntity = permissionRepository.findById(code);
+        if (permissionEntity.isPresent() && permissionEntity.get().getCode() != null)
+            throw new UserException("exception.uniquePermissionCode");
+    }
+
+    /**
+     * Валидация на удаление права доступа
+     * Запрещено удалять право доступа, если существует роль или система с данным правом доступа
+     */
+    private void checkRolesWithSuchPermission(PermissionEntity permissionEntity) {
+        if (!permissionEntity.getRoleList().isEmpty())
+            throw new UserException("exception.roleOrSystemWithSuchPermissionExists");
+    }
+
+    /**
+     * Право доступа не можеть быть родителем самого себя
+     * и не может быть родителем родительского права доступа
+     */
+    private void checkParent(PermissionUpdateForm permission) {
+        if (permission.getParent() != null && permission.getParent().getCode() != null) {
+            if (permission.getCode().equals(permission.getParent().getCode()))
+                throw new UserException("exception.selfParent");
+
+            PermissionEntity parentPermission = permissionRepository.findById(permission.getParent().getCode()).get();
+            while (parentPermission != null) {
+                if (permission.getCode().equals(parentPermission.getCode()))
+                    throw new UserException("exception.cyclicParent");
+                parentPermission = parentPermission.getParentPermission();
+            }
+        }
     }
 }
